% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{asJSONS4}
\alias{asJSONS4}
\title{Convert an Object to a Dictionary or Array in JSON Notation}
\usage{
asJSONS4(object, prototype, exclude = character(), level = 0)
}
\arguments{
\item{object}{The object to convert.}

\item{prototype}{The prototype object (supplied from the evaluator).}

\item{exclude}{Slots to exclude from the dictionary.}

\item{level}{The level of expansion of objects within the original object.}
}
\description{
The general converter for an object with a formal class
or an object to be described via its S3 class, data and attributes.
Not usually called directly, but from \code{\link{objectAsJSON}} or method in an interface package.
}
\references{
Chambers, John M. (2016)
\emph{Extending R},
Chapman & Hall/CRC.
( Chapter 13, discussing this package, is included in the package: \url{../doc/Chapter_XR.pdf}.)
}
