% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Interface.R
\docType{methods}
\name{asRObject}
\alias{asRObject}
\alias{asRObject,ProxyObject-method}
\alias{asRObject,list-method}
\alias{asRObject,vector_R-method}
\title{Specialize the R Object Returned from an Interface Evaluator}
\usage{
asRObject(object, evaluator)

\S4method{asRObject}{ProxyObject}(object, evaluator)

\S4method{asRObject}{vector_R}(object, evaluator)

\S4method{asRObject}{list}(object, evaluator)
}
\description{
The result of a server language expression is returned as a string, using the JSON
standard notation to represent a scalar, list or dictionary.  Methods for this function
get the simple R object obtained from deparsing and interpret it generally.
}
\details{
The methods supplied with the `XR` pacakage handle the standard mechanisms for interpretation.
Additional methods are likely to interpret proxy class objects for which the standard
XR representation in terms of class and slots is not what's actually wanted.
}
\section{Methods (by class)}{
\itemize{
\item \code{ProxyObject}: When a proxy object appears, usually as an element of a list, it
is expanded, by using the `Get()` method of the evaluator and calling `asRObject()` on the
result.

\item \code{vector_R}: To distinguish typed R vectors from a general JSON list, encode the
desired data as an object from the "vector_R" class.

\item \code{list}: Both lists and dictionaries will come here from the basic conversion.
`names(object)` will either be NULL or all non-empty, from a dictionary.
}}

