% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R, R/ProxyClass.R
\docType{class}
\name{AssignedProxy-class}
\alias{$,AssignedProxy-method}
\alias{$<-,AssignedProxy-method}
\alias{AssignedProxy}
\alias{AssignedProxy-class}
\alias{show,AssignedProxy-method}
\title{Class for Assigned Proxy Objects and Related Mechanisms}
\usage{
\S4method{show}{AssignedProxy}(object)

\S4method{$}{AssignedProxy}(x, name)

\S4method{$}{AssignedProxy}(x, name) <- value
}
\description{
The `AssignedProxy` class is used by interface packages to return a reference
to an assigned server object.  The \R user can then supply this object anywhere
in later interface computations, just as one would use the name of an \R object
in a function call or other expression.
}
\details{
The virtual class `ProxyObject` is a superclass, designed to allow
other mechanisms for proxy objects (none exists at this time).
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Prints the language, server class and optional module and size

\item \code{$}: Signals error; must be defined for server langauge

\item \code{$<-}: Signals error; must be defined for server langauge
}}
\section{Slots}{

\describe{
\item{\code{.Data}}{The `AssignedProxy` class is a subclass of `character`; the actual character string
will be generated by the interface and is unique over the session, so long as the `XR`
package stays loaded.}

\item{\code{serverClass}}{The server language class and module for the corresponding object.}

\item{\code{size}}{The size (usually, length) of the server object, if that makes sense.
Can be used to make decisions about handling large objects.}

\item{\code{evaluator}}{The evaluator object that returned the proxy.  Having this as a slot allows
interface computations to operate directly on the proxy, without a user-supplied evaluator.}
}}

