% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxyClass.R
\name{dumpProxyClass}
\alias{dumpProxyClass}
\alias{dumpProxyFunction}
\title{Dump Explicit Definition of Proxy Classes and Functions}
\usage{
dumpProxyClass(file = .dumpFileName(Class), ...,
  where = topenv(parent.frame), objName = Class,
  doSet = XR::setProxyClass)

dumpProxyFunction(file = .dumpFileName(fname, "Function"), ...,
  where = topenv(parent.frame), objName = obj@name,
  doSet = XR::ProxyFunction)
}
\arguments{
\item{file}{where to write the generated R code.  By default makes up a name from the function or  class names.}

\item{...}{the arguments that would be given to setProxyClass() or setProxyFunction()}

\item{where}{environment for the function or class definitions; the same default as for setClass, etc..}

\item{objName}{the name to use when assigning the class generator or the proxy function object; defaults respectively to
the R class name and the server language function name.}

\item{doSet}{the function to create the object; defaults to XR:setProxyClass
and XR::ProxyFunction.  Individual server language interfaces will supply doSet as their customized versions of
those two functions.}
}
\description{
To avoid getting server language class and function information at load time,
which won't work with narrow-minded
package repositories such as CRAN, these functions generate R code to define the proxy classes
and functions explicitly.  The calls to the functions will usually be the same as the calls to \code{\link{setProxyClass}}
or \code{\link{ProxyFunction}}, plus a first argument being the file or open connection to which the output will be sent.
}
\details{
The package creator should call \code{dumpProxyFunction()} and/or \code{dumpProxyClass()} to write
files of R source.  The files go into the application package's source, defining the same
proxy functions or classes explicitly.
}
\section{Functions}{
\itemize{
\item \code{dumpProxyFunction}: Dumps the constructed proxy function to the specified file or connection
}}

