% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ProxyClass.R
\name{dumpProxyClasses}
\alias{dumpProxyClasses}
\alias{dumpProxyFunctions}
\title{Dump Explicit Definition of Proxy Classes and Functions}
\usage{
dumpProxyClasses(Classes, file = .dumpFileName(Classes), where = .GlobalEnv)

dumpProxyFunctions(functions, file = .dumpFileName(functions, "Function"),
  where = topenv(parent.frame()))
}
\arguments{
\item{Classes}{the names of the server language classes for which proxy classes should have been created.}

\item{file}{where to write the generated R code.  By default makes up a name from the function or  class names.}

\item{where}{where to find the function or class definitions; by deafault and usually the global environment.}

\item{functions}{the names of the server language functions for which proxies have been created.}
}
\description{
To avoid getting server language class and function information at load time,
which won't work with narrow-minded
package repositories such as CRAN, these functions generate R code to define the proxy classes
and functions explicitly.
}
\details{
The package creator should create the desired proxy functions and/or classes from an interactive session.
Then call \code{dumpProxyFunctions()} and/or \code{dumpProxyClasses()} to write
files of R source.  The files go into the application package's source, defining the same
proxy functions or classes explicitly.
}
\section{Functions}{
\itemize{
\item \code{dumpProxyFunctions}: 
}}

