% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\docType{methods}
\name{evaluatorAction}
\alias{evaluatorAction}
\alias{evaluatorAction,language-method}
\alias{evaluatorAction,pathEl-method}
\title{Carry Out an Evaluator Initialization Action}
\usage{
evaluatorAction(action, ev)

\S4method{evaluatorAction}{language}(action, ev)

\S4method{evaluatorAction}{pathEl}(action, ev)
}
\arguments{
\item{action}{the action from the table.  Must be an expression or some special
class, typically a path element to add to the server path.}

\item{ev}{the evaluator.}
}
\description{
This function is called from the Startup() method of an evalautor and is not
useful to be called directly.  It is exported to make it visible from within
a subclass of "Interface".
}
\section{Methods (by class)}{
\itemize{
\item \code{language}: a language object, just evaluate it.

\item \code{pathEl}: a "pathEl" object to add to the server search path.
}}

