% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{fillNames}
\alias{fillNames}
\title{Utilities for Server-Language Specific Use}
\usage{
fillNames(object, noNamesOK = FALSE)
}
\arguments{
\item{object}{a list object, possibly with empty or duplicate names.}

\item{noNamesOK}{what to do with a list having no names---leave it alone or fill them all in?}
}
\description{
A utility to fill in blank names in list elements to make it valid for a dictionary.
}
\references{
Chambers, John M. (2016)
\emph{Extending R},
Chapman & Hall/CRC.
( Chapter 13, discussing this package, is included in the package: \url{../doc/Chapter_XR.pdf}.)
}
