% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Interface.R
\name{fixHelpTopic}
\alias{fixHelpTopic}
\title{Make a Help Topic an Explicit Character String}
\usage{
fixHelpTopic(topic)
}
\arguments{
\item{topic}{The \emph{expression} supplied by your user.  If this was a name, it will be
taken literally, otherwise evaluated two levels up the call stack, which should return a character string.}
}
\description{
A helper function to pass on a help topic (specifically for a reference-class method) to
another help-style function; i.e., so the user could have supplied either a name or a
general expression that evaluates to a character string.
}
\references{
Chambers, John M. (2016)
\emph{Extending R},
Chapman & Hall/CRC.
( Chapter 12, discussing this package, is included in the package: \url{../doc/Chapter_XR.pdf}.)
}

