% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Interface.R
\name{getInterface}
\alias{getInterface}
\title{Get or start an evaluator for an interface}
\usage{
getInterface(Class, ..., new = NA)
}
\arguments{
\item{...}{arguments, if any, are passed to the generator for the evaluator.}

\item{new}{can be used to force or prevent starting a new evaluator.}
}
\value{
an  interface evaluator for this class, starting one if none exists.
}
\description{
Get or start an evaluator for an interface
}
\details{
Specific language interface packages usually supply a convenience function equivalent
to calling \code{getInterface()} for their class; e.g., \code{RPython()} in \code{'XRPython'}

If no \code{Class} is given, the current (i.e., last active) evaluator is returned
}

