% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Interface.R
\docType{class}
\name{InterfaceCondition-class}
\alias{InterfaceCondition-class}
\alias{InterfaceError-class}
\alias{InterfaceWarning-class}
\title{Classes of objects representing errors or other conditions in a server language}
\description{
Errors and warnings generated in evaluating an expression in the server language will
be returned to R as objects from one of these classes.
The interface evaluator will normally throw an error for \code{"InterfaceError"} and
issue a warning for \code{"InterfaceWarning"}.
}

