% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxyClass.R
\name{isProxy}
\alias{isProxy}
\title{Test if an Object is a Proxy}
\usage{
isProxy(object)
}
\arguments{
\item{object}{Any object.}
}
\description{
Returns \code{TRUE} if \code{object} is either a simple proxy or an object from a proxy class.
}
