% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxyClass.R
\name{loadProxyClasses}
\alias{loadProxyClasses}
\title{Set a Load Action for Proxy Classes}
\usage{
loadProxyClasses(classes, module = "", evaluator = XR::getInterface(), ...)
}
\arguments{
\item{classes}{names of one or more classes in the server language}

\item{evaluator}{the evaluator to use, by default the current evaluator.}

\item{...}{arguments to pass on to the setProxyClass() call that will be generated at load tome.}

\item{modules}{modules (aka libraries, packages) in which to find the classes}
}
\description{
Set up a load action to define one or more proxy classes for a package
The call to loadProxyClasses() usually comes in the source for an application package
that uses an interface to one or more server langauges.
The alternative, usually preferred, is to use \code{dumpProxyClass()} to generate an explicit
definition in the source for the package.
}

