% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ProxyClass.R
\docType{methods}
\name{MiscMethods}
\alias{$,AssignedProxy-method}
\alias{$<-,AssignedProxy-method}
\alias{MiscMethods}
\title{Miscellaneous methods}
\usage{
\S4method{$}{AssignedProxy}(x, name)

\S4method{$}{AssignedProxy}(x, name) <- value
}
\arguments{
\item{x,name,value}{Arguments to the operator.}
}
\description{
Convenience methods are provided for operator \code{$} to give more informative error messages
if \code{AssignedProxy} objects are assumed
incorrectly to have a proxy class definition.
}

