% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Interface.R
\name{nameQuote}
\alias{nameQuote}
\title{Plain Double Quote for Names}
\usage{
nameQuote(what)
}
\arguments{
\item{what}{the input string.  Should be a name or something without quotes at least.}
}
\value{
the string with quotes.  But empty strings stay empty.
}
\description{
Utility to surround names of classes, etc with double quotes (not the single quotes of
shQuote or the fancy quotes of dQuote)
}
\references{
Chambers, John M. (2016)
\emph{Extending R},
Chapman & Hall/CRC.
( Chapter 12, discussing this package, is included in the package: \url{../doc/Chapter_XR.pdf}.)
}

