% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Interface.R
\name{noScalar}
\alias{noScalar}
\title{Force an object to be treated as a vector in the server language}
\usage{
noScalar(object)
}
\arguments{
\item{object}{A vector object.  Calling with a non-vector is an error.}
}
\value{
the object, but with the S4 bit turned on.
Relies on the convention that XR interfaces leave S4 objects
as vectors, not scalars, even when they are of length 1
}
\description{
Force an object to be treated as a vector in the server language
}

