% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxyClass.R
\docType{class}
\name{ProxyClass-class}
\alias{ProxyClass-class}
\alias{ProxyClass}
\title{A Class to Describe Classes in the Server Language}
\description{
Used by initialize() methods for proxy class objects and therefore exported for the sake of
packages using the XR model.  Not typically needed by end users.
}
\section{Fields}{

\describe{
\item{\code{ServerClass}}{the name of the corresponding server language class.}

\item{\code{language}}{the name of the server language (locked)}

\item{\code{evaluatorClass}}{the class for an interface evaluator for this proxy class.}
}}


