% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ProxyClass.R
\docType{class}
\name{ProxyClassObject-class}
\alias{ProxyClassObject}
\alias{ProxyClassObject-class}
\title{A Class for Objects that are Proxies for Specific Server Class Objects}
\description{
This class is extended by all specific proxy classes for a particular language.
If a proxy object is returned from the server language whose server class matches a
defined proxy class, then an object from that class is generated.
}
\section{Fields}{

\describe{
\item{\code{.proxyObject}}{the actual proxy reference}

\item{\code{.proxyClass}}{the description of the server language class (name, module, langauge)}

\item{\code{.ev}}{the evaluator that produced this proxy object.}
}}
\references{
Chambers, John M. (2016)
\emph{Extending R},
Chapman & Hall/CRC.
(Chapter 12 is included in the \code{"doc"} directory of this package.)
}

