% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxyClass.R
\name{proxyEvaluator}
\alias{proxyEvaluator}
\title{The Evaluator Function Object Referred to from a Proxy Objec}
\usage{
proxyEvaluator(object)
}
\arguments{
\item{object}{any proxy object}
}
\description{
Any proxy for a server language object contains a reference to the interface evaluator object
used to create the proxy object.  This function retrieves the evaluator (whether or not there is
a proxy class for this object).  The function is called from specialized methods for particular
server langauge classes, as part of a package using the XR package.  End users will not typically
need to call it directly; it is exported to simplify life for the extending package.
}
