% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ProxyClass.R
\docType{class}
\name{ProxyFunction-class}
\alias{ProxyFunction}
\alias{ProxyFunction-class}
\alias{initialize,ProxyFunction-method}
\title{A Class for Proxy Functions}
\usage{
\S4method{initialize}{ProxyFunction}(.Object, name = "", module = "",
  prototype = function(...) NULL, evaluator = getInterface(), ...,
  .get = NA)
}
\description{
A Class for Proxy Functions
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: 
}}
\section{Slots}{

\describe{
\item{\code{name}}{the name of the server language function}

\item{\code{the}}{name of the module, if that needs to be imported}

\item{\code{the}}{class for the evaluator, identifying which server lanaguage is involved.}
}}

