% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R, R/ProxyClass.R
\docType{methods}
\name{proxyName}
\alias{proxyName}
\alias{proxyName,AssignedProxy-method}
\alias{proxyName,ProxyClassObject-method}
\title{Return the Server Language Name Corresponding to a Proxy Object}
\usage{
proxyName(x)

\S4method{proxyName}{AssignedProxy}(x)

\S4method{proxyName}{ProxyClassObject}(x)
}
\arguments{
\item{x}{an object returned from some computation in the server  as a proxy
for that server object.  May be from a proxy class, but doesn't need to be.}
}
\description{
Interface evaluators pass constructed names to the server language evaluator, arranged
to be unique within and between evaluators.  This function returns the name when given a
proxy object.
}
\section{Methods (by class)}{
\itemize{
\item \code{AssignedProxy}: for this class, the name is the object (which extends class
"character")

\item \code{ProxyClassObject}: this class has a proxy object as a field.
}}

