% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\name{Scalar}
\alias{Scalar}
\alias{Scalar-class}
\alias{integerScalar-class}
\alias{logicalScalar-class}
\alias{numericScalar-class}
\title{Scalar objects: declarations and classes}
\usage{
Scalar(object)
}
\arguments{
\item{object}{A valid object for a scalar; that is, a vector of length 1.}
}
\value{
The same value but with a suitable scalar class `numericScalar` etc.
}
\description{
The generator function \code{Scalar} and the corresponding class union indicate to the
interface evaluator that the object should be a valid scalar.  These are not often needed
explicitly, if the default rule converting length-1 vectors to scalars is applied and if
the JSON scalar output is a legal constant in the server language.
}

