% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Interface.R
\name{serverAddToPath}
\alias{serverAddToPath}
\alias{serverImport}
\title{Add to Tables of Search Paths and Import Commands}
\usage{
serverAddToPath(Class, directory,
  package = utils::packageName(topenv(parent.frame())), pos = NA,
  onLoad = NA)

serverImport(Class, ..., package = utils::packageName(topenv(parent.frame())),
  onLoad = NA)
}
\arguments{
\item{Class}{the class of the server-specific evalutor.}

\item{directory}{the directory to add to the search path table.}

\item{package}{the name of the server-specific interface package.}

\item{pos}{where in the list of directories to insert this one.  Defaults to the end.}

\item{onLoad}{is the action being taken at load time.  (May be irrelevant)}

\item{...}{arguments to pass to the evaluator's \code{$ServerImport()} method}
}
\description{
Utilities to add to the tables of search paths and import commands for all evalutors of the
specified class.  Called only from analogous functions in packages for specific languages.
}
\details{
The server-specific information is added to the tables stored by the XR package.  All future
evaluators for the specified interface class will have these directories in their search
path and will import the module information specified.  If a current evaluator for this
class exists, it augments its path or issues the import command, but \emph{previous} evalutors
for this class are not modified.
}

