% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProxyClass.R
\docType{class}
\name{ServerClassDef-class}
\alias{ServerClassDef}
\alias{ServerClassDef-class}
\title{The Definition of a Server Language Class}
\description{
The Definition of a Server Language Class
}
\section{Fields}{

\describe{
\item{\code{fields,methods}}{named lists of the server language fields and methods to be exported}

\item{\code{operators}}{named list of the server class methods that are "operator overloading" of functions.}

\item{\code{readOnly}}{the names of any fields that should be made read-only in the R class}
}}

