% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Interface.R
\name{typeToJSON}
\alias{typeToJSON}
\title{Convert a Simple Object to JSON String}
\usage{
typeToJSON(object, prototype)
}
\arguments{
\item{object}{the object to convert; only its type will be used to select format}

\item{prototype}{the prototype object, supplied from the evaluator calling this function}
}
\description{
Convert a simple object (should have no attributes or formal class) to JSON string.
Called from some specific interface packages, usually to put quotes and escapes into a string.
}
\references{
Chambers, John M. (2016)
\emph{Extending R},
Chapman & Hall/CRC.
( Chapter 12, discussing this package, is included in the package: \url{../doc/Chapter_XR.pdf}.)
}

