% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ProxyClass.R
\name{dumpProxyClass}
\alias{dumpProxyClass}
\alias{dumpProxyFunction}
\title{Dump Explicit Definition of Proxy Classes and Functions}
\usage{
dumpProxyClass(gen, file, Class, objName = Class)

dumpProxyFunction(file, object, objName = object@name)
}
\arguments{
\item{gen}{the generator object for the class.}

\item{file}{where to write the generated R code.  If simply \code{TRUE}, a name is constructed from the function or  class names.
May also be a connection.  If the file or connection is not open, it is opened and then
closed on exit.}

\item{Class}{the server language class name.}

\item{objName}{the name to use when assigning the class generator or the proxy function object; defaults respectively to
the R class name and the server language function name.}

\item{object}{the proxy object, constructed by the initialization method for one of the proxy
function classes, such as \code{"PythonFunction"}.}
}
\description{
To avoid getting server language class and function information at load time,
which may not work with narrow-minded
package repositories such as CRAN, these functions generate R code to define the proxy classes
and functions explicitly.
Normally not called directly.
}
\details{
The functions will usually be called from  \code{\link{setProxyClass}}
or an initializing method for a subclass of \code{\link{ProxyFunction}},
with an argument for the file or open connection to which the output will be sent.

If called with the working directory set to the package source directory,
the files go into the application package's R source, defining the same
proxy functions or classes explicitly.
}
\references{
Chambers, John M. (2016)
\emph{Extending R},
Chapman & Hall/CRC.
(Chapter 12 is included in the \code{"doc"} directory of this package.)
}

