% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\docType{class}
\name{ProxyObject-class}
\alias{ProxyObject-class}
\title{Class Union to Represent Proxy Objects}
\description{
A virtual class to include all classes that can represent proxy objects
in the server language.  May be extended by the interface for a particular language.
}
