% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ProxyClass.R
\docType{class}
\name{ServerClassDef-class}
\alias{ServerClassDef}
\alias{ServerClassDef-class}
\title{The Definition of a Server Language Class}
\description{
The Definition of a Server Language Class
}
\section{Fields}{

\describe{
\item{\code{fields,methods}}{named lists of the server language fields and methods to be exported}

\item{\code{operators}}{named list of the server class methods that are "operator overloading" of functions.}

\item{\code{readOnly}}{the names of any fields that should be made read-only in the R class}
}}
\references{
Chambers, John M. (2016)
\emph{Extending R},
Chapman & Hall/CRC.
( Chapter 12, discussing this package, is included in the package: \url{../doc/Chapter_XR.pdf}.)
}

