% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Interface.R
\docType{class}
\name{vector_R-class}
\alias{initialize,vector_R-method}
\alias{vectorR}
\alias{vector_R-class}
\title{A class that facilitates returning R vectors via a list in JSON}
\usage{
\S4method{initialize}{vector_R}(.Object, data,
  type = typeToClass(typeof(.Object@data)), missing)
}
\description{
Server language code will return a dictionary in which data= is a JSON-style list
and type= is the R vector type desired.  See the asRObject() method described below.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: Expects usually to get only the `data` argument.
}}
\section{Slots}{

\describe{
\item{\code{data}}{The actual vector data.}

\item{\code{type}}{The string for the \R type intended. In spite of the slot name, this really the class;
for example, "numeric" rather than "double".}

\item{\code{missing}}{The index of NA's in this vector.  Needed because most server languages only
treat `NaN` for doubles and have no mechanism for `NA` in other types.}
}}

